requirements.txt
tushare
BeautifulSoup4
html5lib
TA-Lib
schedule
lxml
requests
gevent
pytesseract
pandas
pillow
logbook
